# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AdminMeAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.AdminMeBody'
    type: object
  models.AdminMeBody:
    properties:
      role:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  models.Book:
    properties:
      author:
        type: string
      created_at:
        type: string
      id:
        type: integer
      owner_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.CreateBook:
    properties:
      author:
        type: string
      title:
        type: string
    required:
    - author
    - title
    type: object
  models.HealthOKBody:
    properties:
      data:
        type: string
    type: object
  models.LoginAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.LoginTokenBody'
    type: object
  models.LoginTokenBody:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
    type: object
  models.LoginUser:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LogoutAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.LogoutSuccessBody'
    type: object
  models.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  models.LogoutSuccessBody:
    properties:
      message:
        type: string
    type: object
  models.PatchBook:
    properties:
      author:
        type: string
      title:
        type: string
    type: object
  models.RefreshAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.LoginTokenBody'
    type: object
  models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RegisterAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.RegisterSuccessBody'
    type: object
  models.RegisterSuccessBody:
    properties:
      message:
        type: string
    type: object
  models.ReplaceBook:
    properties:
      author:
        type: string
      title:
        type: string
    required:
    - author
    - title
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 'Go/Echo REST API template: books CRUD, register/login/refresh/logout,
    Redis-backed list cache and optional JWT denylist, Postgres via GORM, Mongo access
    logs, rate limiting, and Swagger.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: golang-rest-api-template
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: Health payload in standard envelope
          schema:
            $ref: '#/definitions/models.HealthOKBody'
      summary: ping example
      tags:
      - example
  /admin/me:
    get:
      description: Returns the authenticated admin's username, user id, and role from
        JWT claims. Example admin-only route for RBAC.
      produces:
      - application/json
      responses:
        "200":
          description: Admin identity in standard envelope
          schema:
            $ref: '#/definitions/models.AdminMeAPIResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Current admin identity
      tags:
      - admin
  /books:
    get:
      description: Get a list of books with optional pagination (offset/limit), filters
        (title_like, author_like, owner_id), and sort (sort/order). List entries are
        keyed by a monotonic cache generation (no Redis KEYS) and a canonical query
        after parsing (leading zeros and surrounding whitespace in query params do
        not fragment the cache). Concurrent cache misses for the same query and generation
        are coalesced (singleflight) so only one database read and Redis write runs
        per cache key.
      parameters:
      - default: 0
        description: Offset for pagination (must be >= 0)
        in: query
        name: offset
        type: integer
      - default: 10
        description: Limit for pagination (minimum 1, capped at 100)
        in: query
        name: limit
        type: integer
      - description: Case-insensitive substring filter on title
        in: query
        name: title_like
        type: string
      - description: Case-insensitive substring filter on author
        in: query
        name: author_like
        type: string
      - description: Exact match filter on owner_id
        in: query
        name: owner_id
        type: integer
      - default: id
        description: Sort field (id, title, author, created_at, updated_at, owner_id)
        enum:
        - id
        - title
        - author
        - created_at
        - updated_at
        - owner_id
        in: query
        name: sort
        type: string
      - default: asc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of books
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all books with pagination, filter, and sort
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Create a new book with the given input data
      parameters:
      - description: Create book object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateBook'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    delete:
      description: Delete the book with the given ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted book
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: book not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Delete a book by ID
      tags:
      - books
    get:
      description: Get details of a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved book
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Book not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Find a book by ID
      tags:
      - books
    patch:
      consumes:
      - application/json
      description: Updates only fields present in the JSON body (at least one of title
        or author).
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.PatchBook'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: book not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Partially update a book by ID (PATCH)
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Replaces both title and author. Use PATCH for partial updates.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Full book fields
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ReplaceBook'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: book not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Replace a book by ID (PUT)
      tags:
      - books
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using username and password; returns access
        JWT and opaque refresh token
      parameters:
      - description: User login object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens in standard envelope
          schema:
            $ref: '#/definitions/models.LoginAPIResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Authenticate a user
      tags:
      - user
  /logout:
    post:
      consumes:
      - application/json
      description: Revokes refresh token(s) for the authenticated user. Empty body
        revokes all refresh sessions and sets a per-user access-token revoke_before
        cutoff (when Redis denylist is enabled) so other devices' access JWTs are
        rejected immediately. When refresh_token is provided, only that token family
        is revoked. The current access JWT jti is always denylisted when denylist
        is enabled.
      parameters:
      - description: Optional refresh token to revoke a single family; omit to logout
          all sessions
        in: body
        name: body
        schema:
          $ref: '#/definitions/models.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout confirmation
          schema:
            $ref: '#/definitions/models.LogoutAPIResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Log out and revoke tokens
      tags:
      - user
  /refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access JWT and rotated
        refresh token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens in standard envelope
          schema:
            $ref: '#/definitions/models.RefreshAPIResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Refresh access token
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the given username and password
      parameters:
      - description: User registration object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "201":
          description: Registration message in standard envelope
          schema:
            $ref: '#/definitions/models.RegisterAPIResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Register a new user
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  JwtAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"